pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "isilon-mcp-server"
version = "1.0.0"
description = "Model Context Protocol (MCP) server for Dell PowerScale (Isilon) OneFS API integration"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Isilon MCP Server Contributors" }
]
maintainers = [
    { name = "Isilon MCP Server Contributors" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "isilon",
    "powerscale",
    "dell",
    "onefs",
    "storage",
    "nas",
    "api",
    "automation",
    "llm",
    "ai",
    "claude"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Filesystems",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
http = [
    "starlette>=0.32.0",
    "uvicorn>=0.24.0",
    "sse-starlette>=1.6.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "respx>=0.20.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "isilon-mcp-server[http,dev,docs]",
]

[project.scripts]
isilon-mcp = "isilon_mcp.main:run"

[project.urls]
Homepage = "https://github.com/sachdev27/isilon-mcp-server"
Documentation = "https://github.com/sachdev27/isilon-mcp-server#readme"
Repository = "https://github.com/sachdev27/isilon-mcp-server.git"
Issues = "https://github.com/sachdev27/isilon-mcp-server/issues"
Changelog = "https://github.com/sachdev27/isilon-mcp-server/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["isilon_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/isilon_mcp",
    "/powerscale_9.7_comprehensive_openapi.json",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["isilon_mcp"]
skip_gitignore = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument (for ASGI handlers and test fixtures)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "ARG002"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["isilon_mcp"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
